Refactor watch components and enhance shopping cart functionality
- Removed HeroSection component and replaced it with ProductHeroSection.
- Updated Products component to include filtering and sorting options for watches.
- Deleted WatchCard component as it was no longer needed.
- Enhanced WatchData with additional watch entries and categories.
- Updated routing to reflect changes in cart and product pages.
- Added AboutSection and ContactSection to relevant pages.
- Implemented ShoppingCart component with increment, decrement, and remove item functionality.
- Improved UI for shopping cart and product detail pages.
