+++ src/components/FeatureSection.jsx
@@ -1,45 +1,107 @@
-import React from 'react';
-import { useNavigate } from 'react-router-dom';
+import React, { useRef, useState } from 'react';
+import { Link, useNavigate } from 'react-router-dom';
 import watchData from "./WatchData";
 
 
 function FeatureSection() {
-  const navigate = useNavigate();
+  const [showAll, setShowAll] = useState(false);
+    const [selectedImage, setSelectedImage] = useState(null);
+    const gridRef = useRef(null);
+    const navigate = useNavigate();
+  
+    const handleToggle = () => {
+      if (showAll) {
+        gridRef.current.scrollIntoView({ behavior: 'smooth' });
+      }
+      setShowAll(!showAll);
+    };
+  
+    // Hardcoded add to cart
+    const handleAddToCart = () => {
+      alert('Added to cart');
+      navigate('/ShoppingCart');
+    };
   const featuredWatches = watchData.slice(0, 6); 
 
   return (
-    <section className="bg-white py-10 px-6 text-gray-800 font-sans">
-      <h2 className="text-2xl font-bold mb-6 text-center">FEATURED PRODUCTS</h2>
+    <section className="bg-white text-gray-800 px-6 py-10 font-sans">
+       <h1 className="text-center text-3xl font-bold mb-12">FEATURED PRODUCTS</h1>
+        <div
+          ref={gridRef}
+          className="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 gap-6 max-w-6xl mx-auto"
+        >
+          {featuredWatches.map((item, index) => (
+            <div
+              key={item.id}
+              className="bg-gray-100 border p-4 rounded-md shadow-md hover:shadow-xl hover:scale-105 transform transition duration-300 ease-in-out cursor-pointer"
+              style={{
+                animationDelay: `${index * 0.1}s`,
+                animationFillMode: 'forwards',
+              }}
+            >
+              <img
+                src={item.image}
+                alt={item.name}
+                onClick={() => setSelectedImage(item.image)}
+                className="w-full h-40 object-contain mb-3 rounded"
+              />
 
-      <div className="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 gap-6 max-w-6xl mx-auto">
-        {featuredWatches.map((item) => (
+              <div className="flex justify-between items-center mb-2">
+                <div className="bg-pink-800 text-white text-xs font-bold px-4 py-2 rounded">
+                  30% OFF
+                </div>
+
+                {/* Cart icon to auto navigate with hardcoded item */}
+                <div
+                  className="text-gray-700 hover:text-pink-800 cursor-pointer"
+                  onClick={handleAddToCart}
+                >
+                  <svg
+                    xmlns="http://www.w3.org/2000/svg"
+                    className="w-6 h-6"
+                    fill="none"
+                    viewBox="0 0 24 24"
+                    stroke="currentColor"
+                  >
+                    <path
+                      strokeLinecap="round"
+                      strokeLinejoin="round"
+                      strokeWidth="2"
+                      d="M16 11V9a4 4 0 00-8 0v2M5 8h14l1 12H4L5 8z"
+                    />
+                  </svg>
+                </div>
+              </div>
+
+              <h3 className="text-sm font-bold line-clamp-2 h-9">{item.name}</h3>
+              <p className="text-xs font-semibold text-amber-500 mt-1">
+                PRICE <span className="font-bold ml-1">{item.price}</span>
+              </p>
+            </div>
+          ))}
+        </div>
+
+        {/* Image Zoom Modal */}
+        {selectedImage && (
           <div
-            key={item.id}
-            className="bg-gray-100 border p-4 rounded-md shadow-md hover:shadow-xl transform hover:scale-105 transition duration-300 ease-in-out cursor-pointer"
+            className="fixed inset-0 bg-black bg-opacity-80 flex items-center justify-center z-50"
+            onClick={() => setSelectedImage(null)}
           >
             <img
-              src={item.image}
-              alt={item.name}
-              className="w-full h-40 object-contain mb-3 rounded"
+              src={selectedImage}
+              alt="Zoomed Watch"
+              className="max-w-full max-h-[90vh] rounded-lg transition duration-300"
             />
-            <h3 className="text-sm font-bold line-clamp-2 h-9">{item.name}</h3>
-            <p className="text-xs font-semibold text-amber-500 mt-1">
-              PRICE <span className="font-bold ml-1">{item.price}</span>
-            </p>
           </div>
-        ))}
-      </div>
+        )}
 
-      
-      <div className="text-center mt-8">
-        <button
-          onClick={() => navigate('/products')}
-          className="bg-pink-800 text-white px-6 py-2 rounded hover:bg-gray-900 transition"
-        >
-          VIEW MORE PRODUCTS
-        </button>
+        {/* View More Button */}
+      <div className="text-center mt-10">
+        <Link to="/watches" className="bg-red-900 text-white px-6 py-2 rounded hover:bg-gray-900 transition cursor-pointer">
+          VIEW MORE
+        </Link>
       </div>
-    </section>
+      </section>
   );
 }
