+++ src/components/Footer.jsx
@@ -2,28 +2,29 @@ import Stripe from '/Home/StripeLogo.png';
 import paypal from '/Home/Paypal.png';
 import Visa from '/Home/VisaLogo.png';
 import MasterCard from '/Home/masterCardLogo.png';
+import { Link } from 'react-router-dom';
 
 export default function Footer({ logo }) {
     return (
-        <footer className="bg-[#622525B2] text-white pt-12 pb-6 sm:px-6 px-4">
+        <footer className="bg-[#622525B2] text-white pt-12 pb-6">
             <div className="max-w-4xl mx-auto grid grid-cols-2 md:grid-cols-3 gap-6">
                 <div className="flex flex-col items-start">
                     <h3 className="font-extrabold text-xl mb-2">Legal</h3>
-                    <ul className="space-y-1 text-sm font-semibold">
+                    <ul className="space-y-2 text-sm font-semibold">
                         <li><a href="">Privacy policy</a></li>
                         <li><a href="">Shipping policy</a></li>
                         <li><a href="">Terms and Condition</a></li>
                         <li><a href="">Refund policy</a></li>
-                        <li><a href="">Contact us</a></li>
                     </ul>
                 </div>
                 <div className="flex flex-col items-start">
                     <h3 className="font-extrabold text-xl mb-2">Menu</h3>
                     <ul className="space-y-1 text-sm font-semibold">
-                        <li><a href="" className=''>Home</a></li>
-                        <li><a href="" className=''>Sale</a></li>
-                        <li><a href="" className=''>About us</a></li>
-                        <li><a href="" className=''>Watches</a></li>
+                        <li><Link to="/" className=''>Home</Link></li>
+                        <li><Link to="/watches" className=''>Sale</Link></li>
+                        <li><Link to="/about" className=''>About us</Link></li>
+                        <li><Link to="/watches" className=''>Watches</Link></li>
+                        <li><Link to="/contact" className=''>Contact Us</Link></li>
                     </ul>
                 </div>
                 <div className="flex flex-col items-center justify-center">
@@ -32,7 +33,7 @@ export default function Footer({ logo }) {
             </div>
             <div className='max-w-4xl mx-auto text-center'>
                 <div className=" flex items-center justify-center space-x-4">
-                    <img src={Stripe} alt="Stripe Logo" className="md:w-24 w-20 h-auto " />
+                    <a href=""><img src={Stripe} alt="Stripe Logo" className="md:w-24 w-20 h-auto " /></a>
                     <a href="" className=''><img src={paypal} alt="PayPal" className="w-24 h-auto rounded-xl py-3 px-2  bg-blue-900" /></a>
                     <a href=""><img src={Visa} alt="Visa" className="w-24 h-auto" /></a>
                     <a href=""><img src={MasterCard} alt="MasterCard" className="w-24 h-auto rounded-xl py-1 px-2 bg-gray-800" /></a>
