+++ src/components/WatchData.js
@@ -3,27 +3,41 @@ import watch2 from '/WatchProduct/watch 2.png';
 import watch3 from '/WatchProduct/watch 3.png';
 import watch4 from '/WatchProduct/watch 1.png';
 import watch5 from '/WatchProduct/watch 2.png';
-import watch6 from '/WatchProduct/watch 3.png';
-import watch7 from '/WatchProduct/watch 1.png';
-import watch8 from '/WatchProduct/watch 2.png';
-import watch9 from '/WatchProduct/watch 3.png';
-import watch10 from '/WatchProduct/watch 1.png';
-import watch11 from '/WatchProduct/watch 2.png';
-import watch12 from '/WatchProduct/watch 3.png';
 
+// const watchData = [
+//   { id: 1, name: 'QUADRO CRYSTAL WATCH IN GOLD', price: '$677', image: watch1 },
+//   { id: 2, name: 'QUADRO CRYSTAL WATCH IN BLACK', price: '$899', image: watch2 },
+//   { id: 3, name: 'QUADRO CRYSTAL WATCH IN GOLD', price: '$677', image: watch3 },
+//   { id: 4, name: 'QUADRO CRYSTAL WATCH IN GOLD', price: '$677', image: watch4 },
+//   { id: 5, name: 'QUADRO CRYSTAL WATCH IN BLACK', price: '$899', image: watch5 },
+//   { id: 6, name: 'QUADRO CRYSTAL WATCH IN GOLD', price: '$677', image: watch6 },
+//   { id: 7, name: 'QUADRO CRYSTAL WATCH IN GOLD', price: '$677', image: watch7 },
+//   { id: 8, name: 'QUADRO CRYSTAL WATCH IN BLACK', price: '$899', image: watch8 },
+//   { id: 9, name: 'QUADRO CRYSTAL WATCH IN GOLD', price: '$677', image: watch9 },
+//   { id: 10, name: 'QUADRO CRYSTAL WATCH IN GOLD', price: '$677', image: watch10 },
+//   { id: 11, name: 'QUADRO CRYSTAL WATCH IN BLACK', price: '$899', image: watch11 },
+//   { id: 12, name: 'QUADRO CRYSTAL WATCH IN GOLD', price: '$677', image: watch12, category: "Men Watches" },
+// ];
 const watchData = [
-  { id: 1, name: 'QUADRO CRYSTAL WATCH IN GOLD', price: '$677', image: watch1 },
-  { id: 2, name: 'QUADRO CRYSTAL WATCH IN BLACK', price: '$899', image: watch2 },
-  { id: 3, name: 'QUADRO CRYSTAL WATCH IN GOLD', price: '$677', image: watch3 },
-  { id: 4, name: 'QUADRO CRYSTAL WATCH IN GOLD', price: '$677', image: watch4 },
-  { id: 5, name: 'QUADRO CRYSTAL WATCH IN BLACK', price: '$899', image: watch5 },
-  { id: 6, name: 'QUADRO CRYSTAL WATCH IN GOLD', price: '$677', image: watch6 },
-  { id: 7, name: 'QUADRO CRYSTAL WATCH IN GOLD', price: '$677', image: watch7 },
-  { id: 8, name: 'QUADRO CRYSTAL WATCH IN BLACK', price: '$899', image: watch8 },
-  { id: 9, name: 'QUADRO CRYSTAL WATCH IN GOLD', price: '$677', image: watch9 },
-  { id: 10, name: 'QUADRO CRYSTAL WATCH IN GOLD', price: '$677', image: watch10 },
-  { id: 11, name: 'QUADRO CRYSTAL WATCH IN BLACK', price: '$899', image: watch11 },
-  { id: 12, name: 'QUADRO CRYSTAL WATCH IN GOLD', price: '$677', image: watch12, category: "Men Watches" },
+  { id: 1, name: 'Quardro Crystal Watch in Gold', price: 677, category: 'Men', image: watch1, discount: 0 },
+  { id: 2, name: 'Quardro Crystal Watch in Black', price: 899, category: 'Women', image: watch2, discount: 0 },
+  { id: 3, name: 'Quardro Crystal Watch in Gold', price: 677, category: 'Sale', image: watch3, discount: 30 },
+  { id: 4, name: 'Quardro Crystal Watch in Gold', price: 677, category: 'Sale', image: watch4, discount: 40 },
+  { id: 5, name: 'Quardro Crystal Watch in Black', price: 899, category: 'Women', image: watch5, discount: 0 },
+  { id: 6, name: 'Quardro Crystal Watch in Gold', price: 677, category: 'Sale', image: watch3, discount: 30 },
+  { id: 7, name: 'Quardro Crystal Watch in Gold', price: 677, category: 'Sale', image: watch2, discount: 20 },
+  { id: 8, name: 'Quardro Crystal Watch in Black', price: 899, category: 'Men', image: watch5, discount: 0 },
+  { id: 9, name: 'Quardro Crystal Watch in Gold', price: 677, category: 'Sale', image: watch1, discount: 10 },
+  { id: 10, name: 'Quardro Crystal Watch in Gold', price: 677, category: 'Sale', image: watch5, discount: 50 },
+  { id: 11, name: 'Quardro Crystal Watch in Black', price: 899, category: 'Men', image: watch1, discount: 0 },
+  { id: 12, name: 'Quardro Crystal Watch in Gold', price: 677, category: 'Women', image: watch2, discount: 0 },
+  { id: 13, name: 'Quardro Crystal Watch in Gold', price: 677, category: 'Sale', image: watch3, discount: 30 },
+  { id: 14, name: 'Quardro Crystal Watch in Gold', price: 677, category: 'Sale', image: watch2, discount: 20 },
+  { id: 15, name: 'Quardro Crystal Watch in Black', price: 899, category: 'Men', image: watch5, discount: 0 },
+  { id: 16, name: 'Quardro Crystal Watch in Gold', price: 677, category: 'Sale', image: watch1, discount: 10 },
+  { id: 17, name: 'Quardro Crystal Watch in Gold', price: 677, category: 'Sale', image: watch5, discount: 50 },
+  { id: 18, name: 'Quardro Crystal Watch in Black', price: 899, category: 'Men', image: watch1, discount: 0 },
+  { id: 19, name: 'Quardro Crystal Watch in Gold', price: 677, category: 'Women', image: watch2, discount: 0 },
 ];
 
 export default watchData;
